after making final changes, tests remaining
