--- ws/adapter_test.go
@@ -1,4 +1,0 @@
-// Copyright (c) Mainflux
-// SPDX-Licence-Identifier: Apache-2.0
-
-package ws_test
