+++ ws/api/logging.go
@@ -57,7 +57,7 @@ func (lm *loggingMiddleware) Publish(ctx context.Context, thingKey string, msg m
 	return lm.svc.Publish(ctx, thingKey, msg)
 }
 
-func (lm *loggingMiddleware) Subscribe(ctx context.Context, thingKey, chanID, subtopic string, c ws.Client) (err error) {
+func (lm *loggingMiddleware) Subscribe(ctx context.Context, thingKey, chanID, subtopic string, c *ws.Connclient) (err error) {
 	defer func(begin time.Time) {
 		destChannel := chanID
 		if subtopic != "" {
